feat: add user auth and register api

+++ controllers/card.js
@@ -3,23 +3,21 @@ const Card = require('../models/card');
 const {
   DEFAULT_SUCCESS_CODE,
   SUCCESS_CREATED_CODE,
-  NOT_FOUND_ERROR_CODE,
-  DEFAULT_ERROR_CODE,
-  INCORRECT_DATA_ERROR_CODE,
 } = require('../utils/constants');
+const BadRequestError = require('../utils/errors/badRequestError');
+const NotFoundError = require('../utils/errors/notFoundError');
+const ForbiddenError = require('../utils/errors/forbiddenError');
 
-module.exports.getCards = async (req, res) => {
+module.exports.getCards = async (req, res, next) => {
   await Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.status(DEFAULT_SUCCESS_CODE).send(cards))
-    .catch(() => {
-      res.status(DEFAULT_ERROR_CODE).send({
-        message: 'Не удалось получить карточки',
-      });
+    .catch((err) => {
+      next(err);
     });
 };
 
-module.exports.createCard = async (req, res) => {
+module.exports.createCard = async (req, res, next) => {
   const { name, link } = req.body;
 
   await Card.create({
@@ -30,62 +28,50 @@ module.exports.createCard = async (req, res) => {
     .then((card) => res.status(SUCCESS_CREATED_CODE).send(card))
     .catch((err) => {
       if (err instanceof mongoose.Error.ValidationError) {
-        res
-          .status(INCORRECT_DATA_ERROR_CODE)
-          .send({ message: 'Переданы не валидные данные' });
-        return;
+        next(new BadRequestError(err.message));
       }
 
-      res
-        .status(DEFAULT_ERROR_CODE)
-        .send({ message: 'Не удалось создать карточку' });
+      next(err);
     });
 };
 
-module.exports.deleteCard = async (req, res) => {
+module.exports.deleteCard = async (req, res, next) => {
   const { cardId } = req.params;
 
-  await Card.findByIdAndRemove(cardId)
-    .then((card) => {
-      if (!card) {
-        return res
-          .status(NOT_FOUND_ERROR_CODE)
-          .send({ message: 'Карточка не найдена' });
-      }
-      return res.status(DEFAULT_SUCCESS_CODE).send(card);
-    })
-    .catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        res
-          .status(INCORRECT_DATA_ERROR_CODE)
-          .send({ message: 'Переданы не валидные данные' });
-      } else {
-        res
-          .status(DEFAULT_ERROR_CODE)
-          .send({ message: 'Не удалось удалить карточку' });
-      }
+  try {
+    const card = await Card.findById(cardId);
+    if (!card) {
+      return new NotFoundError('Карточка не найдена');
+    }
+    if (card.owner.toString() !== cardId) {
+      return new ForbiddenError('Нельзя удалять чужие карточки');
+    }
+    await Card.findByIdAndRemove(cardId);
+    res.send({
+      message: 'Карточка удалена',
     });
+  } catch (err) {
+    if (err instanceof mongoose.Error.CastError) {
+      next(new BadRequestError('Переданы не валидные данные'));
+    } else {
+      next(err);
+    }
+  }
 };
 
-const updateLikes = async (req, res, update) => {
+const updateLikes = async (req, res, next, update) => {
   await Card.findByIdAndUpdate(req.params.cardId, update, { new: true })
     .then((card) => {
       if (!card) {
-        return res
-          .status(NOT_FOUND_ERROR_CODE)
-          .send({ message: 'Карточка не найдена' });
+        return new NotFoundError('Карточка не найдена');
       }
       return res.status(DEFAULT_SUCCESS_CODE).send(card);
     })
     .catch((err) => {
       if (err instanceof mongoose.Error.CastError) {
-        return res
-          .status(INCORRECT_DATA_ERROR_CODE)
-          .send({ message: 'Переданы не валидные данные' });
+        next(new BadRequestError('Переданы не валидные данные'));
       }
-      return res
-        .status(DEFAULT_ERROR_CODE)
-        .send({ message: 'Не удалось изменить карточку' });
+      next(err);
     });
 };
